Don't iterate 10 times, but huge morphological filter

+++ igvc_navigation/src/mapper/map_utils.h
@@ -57,7 +57,7 @@ struct GroundFilterOptions
 struct RemoveOccupiedOptions
 {
   bool enable;
-  double kernel_size;
+  int kernel_size;
 };
 
 struct GroundPlane
